sagemaker/code/requirements.txt
boto3
sagemaker
snowflake-connector-python